# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nike Pillay Inc Law Firm API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface AdminLoginRequest {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
}

export interface ChangePasswordRequest {
  currentPassword: string;
  newPassword: string;
}

export interface SiteSettings {
  id: number;
  firmName?: string;
  tagline?: string;
  heroHeading?: string;
  heroSubheading?: string;
  email?: string;
  phone?: string;
  phone2?: string;
  address?: string;
  bbbeeLevel?: string;
  linkedinUrl?: string;
  facebookUrl?: string;
  twitterUrl?: string;
  updatedAt?: string;
}

export interface UpdateSiteSettingsRequest {
  firmName?: string;
  tagline?: string;
  heroHeading?: string;
  heroSubheading?: string;
  email?: string;
  phone?: string;
  phone2?: string;
  address?: string;
  bbbeeLevel?: string;
  linkedinUrl?: string;
  facebookUrl?: string;
  twitterUrl?: string;
}

export interface Service {
  id: number;
  slug: string;
  title: string;
  summary?: string;
  heroImageUrl?: string;
  howWeAssist?: string;
  typicalMatters?: string[];
  relatedTeamSlugs?: string[];
  sortOrder?: number;
  isPublished?: boolean;
  createdAt?: string;
  updatedAt?: string;
}

export interface UpsertServiceRequest {
  slug: string;
  title: string;
  summary?: string;
  heroImageUrl?: string;
  howWeAssist?: string;
  typicalMatters?: string[];
  relatedTeamSlugs?: string[];
  sortOrder?: number;
  isPublished?: boolean;
}

export interface Person {
  id: number;
  slug: string;
  firstName: string;
  lastName: string;
  role: string;
  title?: string;
  photoUrl?: string;
  qualifications?: string;
  admissions?: string;
  bio?: string;
  email?: string;
  phone?: string;
  practiceAreas?: string[];
  sortOrder?: number;
  isPublished?: boolean;
  createdAt?: string;
  updatedAt?: string;
}

export interface UpsertPersonRequest {
  slug: string;
  firstName: string;
  lastName: string;
  role: string;
  title?: string;
  photoUrl?: string;
  qualifications?: string;
  admissions?: string;
  bio?: string;
  email?: string;
  phone?: string;
  practiceAreas?: string[];
  sortOrder?: number;
  isPublished?: boolean;
}

export interface Article {
  id: number;
  slug: string;
  title: string;
  category: string;
  summary?: string;
  content?: string;
  imageUrl?: string;
  author?: string;
  publishedAt?: string;
  isPublished?: boolean;
  createdAt?: string;
  updatedAt?: string;
}

export interface UpsertArticleRequest {
  slug: string;
  title: string;
  category: string;
  summary?: string;
  content?: string;
  imageUrl?: string;
  author?: string;
  publishedAt?: string;
  isPublished?: boolean;
}

export interface Event {
  id: number;
  slug: string;
  title: string;
  description?: string;
  location?: string;
  eventDate: string;
  eventEndDate?: string;
  imageUrl?: string;
  registrationUrl?: string;
  isPublished?: boolean;
  createdAt?: string;
  updatedAt?: string;
}

export interface UpsertEventRequest {
  slug: string;
  title: string;
  description?: string;
  location?: string;
  eventDate: string;
  eventEndDate?: string;
  imageUrl?: string;
  registrationUrl?: string;
  isPublished?: boolean;
}

export interface Award {
  id: number;
  title: string;
  awardingBody?: string;
  year?: string;
  description?: string;
  imageUrl?: string;
  sortOrder?: number;
  createdAt?: string;
  updatedAt?: string;
}

export interface UpsertAwardRequest {
  title: string;
  awardingBody?: string;
  year?: string;
  description?: string;
  imageUrl?: string;
  sortOrder?: number;
}

export interface Vacancy {
  id: number;
  slug: string;
  title: string;
  department?: string;
  location?: string;
  type?: string;
  summary?: string;
  description?: string;
  isPublished?: boolean;
  closingDate?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface UpsertVacancyRequest {
  slug: string;
  title: string;
  department?: string;
  location?: string;
  type?: string;
  summary?: string;
  description?: string;
  isPublished?: boolean;
  closingDate?: string;
}

export interface CvSubmissionRequest {
  name: string;
  email: string;
  phone?: string;
  position?: string;
  coverLetter?: string;
}

export interface CvSubmission {
  id: number;
  name: string;
  email: string;
  phone?: string;
  position?: string;
  coverLetter?: string;
  createdAt?: string;
}

export interface Document {
  id: number;
  title: string;
  category?: string;
  fileUrl: string;
  description?: string;
  isPublic?: boolean;
  sortOrder?: number;
  createdAt?: string;
  updatedAt?: string;
}

export interface UpsertDocumentRequest {
  title: string;
  category?: string;
  fileUrl: string;
  description?: string;
  isPublic?: boolean;
  sortOrder?: number;
}

export interface CalculatorRate {
  id: number;
  rateType: string;
  label: string;
  value: string;
  effectiveFrom?: string;
  notes?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface UpsertCalculatorRateRequest {
  rateType: string;
  label: string;
  value: string;
  effectiveFrom?: string;
  notes?: string;
}

export interface ContactRequest {
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  message: string;
}

export interface ContactEnquiry {
  id: number;
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  message: string;
  createdAt?: string;
}

export type ListPeopleParams = {
role?: string;
serviceId?: number;
};

export type ListArticlesParams = {
category?: string;
limit?: number;
offset?: number;
};

export type ListEventsParams = {
upcoming?: boolean;
};