# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nike Pillay Inc Law Firm API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [auth]
      summary: Get current admin user
      responses:
        "200":
          description: Current admin user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ─── Site Settings (public) ──────────────────────────────────────────────────
...
  /articles:
    get:
      operationId: listArticles
      tags: [public]
      summary: List articles
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Articles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
  /articles/{slug}:
    get:
      operationId: getArticle
      tags: [public]
      summary: Get article by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/articles:
    get:
      operationId: adminListArticles
      tags: [admin]
      summary: Admin list articles
      responses:
        "200":
          description: Articles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createArticle
      tags: [admin]
      summary: Create an article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertArticleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/articles/{id}:
    put:
      operationId: updateArticle
      tags: [admin]
      summary: Update an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertArticleRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteArticle
      tags: [admin]
      summary: Delete an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ─── Events ──────────────────────────────────────────────────────────────────
...
  /admin/awards:
    get:
      operationId: adminListAwards
      tags: [admin]
      summary: Admin list awards
      responses:
        "200":
          description: Awards list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Award"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createAward
      tags: [admin]
      summary: Create an award
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertAwardRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Award"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/awards/{id}:
    put:
      operationId: updateAward
      tags: [admin]
      summary: Update an award
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertAwardRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Award"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteAward
      tags: [admin]
      summary: Delete an award
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ─── Vacancies ───────────────────────────────────────────────────────────────
...
  /admin/calculator-rates:
    get:
      operationId: adminListCalculatorRates
      tags: [admin]
      summary: Admin list calculator rates
      responses:
        "200":
          description: Calculator rates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalculatorRate"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createCalculatorRate
      tags: [admin]
      summary: Create a calculator rate entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertCalculatorRateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculatorRate"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/calculator-rates/{id}:
    put:
      operationId: updateCalculatorRate
      tags: [admin]
      summary: Update a calculator rate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertCalculatorRateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculatorRate"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteCalculatorRate
      tags: [admin]
      summary: Delete a calculator rate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ─── Contact ─────────────────────────────────────────────────────────────────
...
  /admin/change-password:
    post:
      operationId: changeAdminPassword
      tags: [admin]
      summary: Change admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    AdminLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
      required: [id, username]
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
...
    Article:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        category:
          type: string
        summary:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        author:
          type: string
        publishedAt:
          type: string
        isPublished:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, slug, title, category]
    UpsertArticleRequest:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        category:
          type: string
        summary:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        author:
          type: string
        publishedAt:
          type: string
        isPublished:
          type: boolean
      required: [slug, title, category]
...
    Award:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        awardingBody:
          type: string
        year:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title]
    UpsertAwardRequest:
      type: object
      properties:
        title:
          type: string
        awardingBody:
          type: string
        year:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer
      required: [title]
...
    CalculatorRate:
      type: object
      properties:
        id:
          type: integer
        rateType:
          type: string
        label:
          type: string
        value:
          type: string
        effectiveFrom:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, rateType, label, value]
    UpsertCalculatorRateRequest:
      type: object
      properties:
        rateType:
          type: string
        label:
          type: string
        value:
          type: string
        effectiveFrom:
          type: string
        notes:
          type: string
      required: [rateType, label, value]
...